move obsolete tests
